requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
jsonschema==4.17.3
pytest==7.3.1
pytest-lazy-fixture==0.6.3
requests==2.31.0
